+++ backend/users/models.py
@@ -1,7 +1,12 @@
 import uuid
+from typing import Literal
+
 from django.contrib.auth.models import AbstractUser, BaseUserManager
 from django.db import models
+from django.core.signals import request_started
+from django.dispatch import receiver
 from django.utils.translation import gettext_lazy as _
+from django.utils import timezone
 
 
 class CustomUserManager(BaseUserManager):
@@ -32,6 +37,21 @@ class CustomUser(AbstractUser):
     homepage = models.URLField(max_length=255, null=True)
     avatar = models.ImageField(upload_to="avatars/", blank=True, null=True)
     avatar_url = models.URLField(blank=True, null=True)
+    expires_at = models.DateTimeField(null=True, blank=True)
+    created_ip = models.GenericIPAddressField(null=True, blank=True)
+    rate_limit = models.CharField(max_length=20, default="10/s")
+
+    def set_rate_limit(self, value: int, unit: Literal["s", "m", "d"]):
+        units = {"s", "m", "d"}  # second, minute, day
+
+        if unit not in units:
+            raise ValueError("Unit must be (s)econd, (m)inute, or (d)ay.")
+
+        if not isinstance(value, int) or value < 0:
+            raise ValueError("Value must be a positive integer.")
+
+        self.rate_limit = f"{value}/{unit}"
+        self.save(update_fields=["rate_limit"])
 
     def save(self, *args, **kwargs):
         # To change User name, change first_name and last_name
@@ -50,9 +70,13 @@ class CustomUser(AbstractUser):
             return self.avatar.url
         if self.avatar_url:
             return self.avatar_url
-        return None
 
     objects = CustomUserManager()
+
+
+@receiver(request_started)
+def delete_expired_users(sender, **kwargs):
+    CustomUser.objects.filter(expires_at__lt=timezone.now()).delete()
 
 
 class OAuthAccount(models.Model):