+++ contrib/scripts/do_dotenv.py
@@ -272,6 +272,22 @@ Path(str(dj_host_data)).mkdir(exist_ok=True, parents=True)
 Path(str(psql_host_conf)).mkdir(exist_ok=True, parents=True)
 Path(str(psql_host_data)).mkdir(exist_ok=True, parents=True)
 
+print("\nDjango Backups:")
+def_backup_dir = def_data_dir / "backups"
+backup_dir_host = var_in(
+    "BACKUP_DIR_HOST",
+    input_text=f"  Backup directory on host [{def_backup_dir}]: ",
+    default=str(def_backup_dir.absolute()),
+)
+Path(str(backup_dir_host)).mkdir(exist_ok=True, parents=True)
+
+sentry_dsn = var_in(
+    "SENTRY_DSN",
+    input_text="  Sentry DSN (leave empty to disable): ",
+    required=False,
+    default="",
+)
+
 print("\nDjango Email config")
 dj_default_from_email = var_in(
     "DEFAULT_FROM_EMAIL",
@@ -369,6 +385,9 @@ variables = {
     ),
     # [Mosqlient]
     "MOSQLIENT_API_URL": "http://localhost:$BACKEND_PORT/api/",
+    # [Backups]
+    "BACKUP_DIR_HOST": str(backup_dir_host),
+    "SENTRY_DSN": sentry_dsn,
 }
 
 if not CI: